cli: add system.settings to debug zip
This is handy to figure out whether a setting was explicitly set.
It may not be as elegant as annotating crdb_internal.cluster_settings,
but it is simple, backwards compatible, and something.

Release note (ops change): cockroach debug zip will now include the raw
system.settings table. This table makes it possible to determine whether
a cluster settings has been explicitly set.