kv: don't leak waiting pushes in MaybeWaitForPush
When a PushTxn enters the TxnWaitQueue, it enters a "waiting pushes" set. This
allows distributed deadlock detection to track who is waiting on who and to form
a dependency graph through which is can detect cycles. Before this commit,
waiting pushes would not be removed from this set until the pushee was
finalized, regardless of whether the push was still waiting or not. This meant
that if the PushTxn's context was canceled, its `waitingPush` would not be
removed from the `waitingPushes` list.

Context cancellation of PushTxn requests is common, especially with recent (as
of v20.1) logic in the lock wait-queues. For instance, `lockTableWaiterImpl`
will often issue a push while also continuing to listen to local lock state
transitions. If it observes one, it cancels the push.

Additionally, the `txnwaitqueue.pusher.waiting` gauge would be incremented when
the PushTxn entered the TxnWaitQueue, but would not be decremented until the
pushee was finalized. This commit fixes that issue as well by keeping all
manipulation of the metric local to `MaybeWaitForPush`, which makes it easier to
reason about.

The effect of this was subtle but significant. I originally noticed that
something was up when running `kv/contention/nodes=4` and looking at the
`txnwaitqueue.pusher.waiting` metric. Around the same time as I saw throughput
stall (10s of qps vs. 100s of qps), I saw the metric shoot up to about 5,000
waiting pushers. This did not make sense, because the test runs 128 concurrent
transactions with an internal concurrency of 4, so there should only ever be up
to 512 pushers in the system at a time, not an order of magnitude more than
that.

During this same period, I saw that the node with the majority of these pushers
was running hotter than other nodes. A CPU profile revealed that the significant
amount of this CPU time was spent in `pendingTxn.getDependentsSet`. This was
another red flag that something was wrong, as this method iterates over all
pushers associated with a pushee txn and collects their transitive dependencies
into a set.

It then became clear that because we were leaking these `waitingPush` objects in
the `waitingPushes` list, a single pusher thread could be responsible for
multiple entries in this list. This led to the list being an order of magnitude
larger than should have been possible, and `pendingTxn.getDependentsSet` being
an order of magnitude more expensive than should have been possible. These calls
were then serializes (see the third commit), leading to significant slowdown in
the time to detect and resolve a transaction deadlock.

This commit fixes this issue by removing `waitingPush` entries from the
`waitingPushes` list on context cancellation or any other early push termination
condition. With this change, we never see the `txnwaitqueue.pusher.waiting`
gauge exceed expected upper bounds, nor do we see `pendingTxn.getDependentsSet`
appear in CPU profiles.

Release note (performance improvement): the performance of transaction deadlock
detection is now more stable even with significant transaction contention.

Release note (bug fix): the `txnwaitqueue.pusher.waiting` metric no longer over-
reports the number of pushing transactions in some cases.